fix: Moving SSH tests to their own build tag
