+++ test/integration_scaffold_test.go
@@ -77,51 +77,6 @@ func TestScaffoldModuleDifferentRevision(t *testing.T) {
 	assert.Contains(t, stderr, "Scaffolding completed")
 }
 
-func TestScaffoldModuleDifferentRevisionAndSsh(t *testing.T) {
-	t.Parallel()
-
-	tmpEnvPath := t.TempDir()
-
-	_, stderr, err := helpers.RunTerragruntCommandWithOutput(t, fmt.Sprintf("terragrunt scaffold --terragrunt-non-interactive --terragrunt-working-dir %s %s --var=Ref=v0.67.4 --var=SourceUrlType=git-ssh", tmpEnvPath, testScaffoldModuleShort))
-	require.NoError(t, err)
-	assert.Contains(t, stderr, "git::ssh://git@github.com/gruntwork-io/terragrunt.git//test/fixtures/inputs?ref=v0.67.4")
-	assert.Contains(t, stderr, "Scaffolding completed")
-}
-
-func TestScaffoldModuleSsh(t *testing.T) {
-	t.Parallel()
-
-	tmpEnvPath := t.TempDir()
-
-	_, stderr, err := helpers.RunTerragruntCommandWithOutput(t, fmt.Sprintf("terragrunt scaffold --terragrunt-non-interactive --terragrunt-working-dir %s %s", tmpEnvPath, testScaffoldModuleGit))
-	require.NoError(t, err)
-	assert.Contains(t, stderr, "Scaffolding completed")
-}
-
-func TestScaffoldModuleTemplate(t *testing.T) {
-	t.Parallel()
-
-	tmpEnvPath := t.TempDir()
-
-	_, stderr, err := helpers.RunTerragruntCommandWithOutput(t, fmt.Sprintf("terragrunt scaffold --terragrunt-non-interactive --terragrunt-working-dir %s %s", tmpEnvPath, testScaffoldTemplateModule))
-	require.NoError(t, err)
-	assert.Contains(t, stderr, "Scaffolding completed")
-	// check that exists file from .boilerplate dir
-	assert.FileExists(t, tmpEnvPath+"/template-file.txt")
-}
-
-func TestScaffoldModuleExternalTemplate(t *testing.T) {
-	t.Parallel()
-
-	tmpEnvPath := t.TempDir()
-
-	_, stderr, err := helpers.RunTerragruntCommandWithOutput(t, fmt.Sprintf("terragrunt scaffold --terragrunt-non-interactive --terragrunt-working-dir %s %s %s", tmpEnvPath, testScaffoldModuleGit, testScaffoldExternalTemplateModule))
-	require.NoError(t, err)
-	assert.Contains(t, stderr, "Scaffolding completed")
-	// check that exists file from external template
-	assert.FileExists(t, tmpEnvPath+"/external-template.txt")
-}
-
 func TestScaffoldErrorNoModuleUrl(t *testing.T) {
 	t.Parallel()
 
@@ -130,29 +85,6 @@ func TestScaffoldErrorNoModuleUrl(t *testing.T) {
 	_, _, err := helpers.RunTerragruntCommandWithOutput(t, "terragrunt scaffold --terragrunt-non-interactive --terragrunt-working-dir "+tmpEnvPath)
 	require.Error(t, err)
 	assert.Contains(t, err.Error(), "No module URL passed")
-}
-
-func TestScaffoldModuleVarFile(t *testing.T) {
-	t.Parallel()
-	// generate var file with specific version, without root include and use GIT/SSH to clone module.
-	varFileContent := `
-Ref: v0.67.4
-EnableRootInclude: false
-SourceUrlType: "git-ssh"
-`
-	varFile := filepath.Join(t.TempDir(), "var-file.yaml")
-	err := os.WriteFile(varFile, []byte(varFileContent), 0644)
-	require.NoError(t, err)
-
-	tmpEnvPath := t.TempDir()
-
-	_, stderr, err := helpers.RunTerragruntCommandWithOutput(t, fmt.Sprintf("terragrunt scaffold --terragrunt-non-interactive --terragrunt-working-dir %s %s --var-file=%s", tmpEnvPath, testScaffoldModuleShort, varFile))
-	require.NoError(t, err)
-	assert.Contains(t, stderr, "git::ssh://git@github.com/gruntwork-io/terragrunt.git//test/fixtures/inputs?ref=v0.67.4")
-	assert.Contains(t, stderr, "Scaffolding completed")
-	content, err := util.ReadFileAsString(tmpEnvPath + "/terragrunt.hcl")
-	require.NoError(t, err)
-	assert.NotContains(t, content, "find_in_parent_folders")
 }
 
 func TestScaffoldLocalModule(t *testing.T) {
@@ -225,24 +157,4 @@ func TestScaffoldWithRootHCL(t *testing.T) {
 	content, err := util.ReadFileAsString(filepath.Join(testPath, "unit", "terragrunt.hcl"))
 	require.NoError(t, err)
 	assert.Contains(t, content, `path = find_in_parent_folders("root.hcl")`)
-}
-
-func TestScaffoldWithCustomDefaultTemplate(t *testing.T) {
-	t.Parallel()
-
-	tmpEnvPath := helpers.CopyEnvironment(t, testScaffoldWithCustomDefaultTemplate)
-	helpers.CleanupTerraformFolder(t, tmpEnvPath)
-	testPath := util.JoinPath(tmpEnvPath, testScaffoldWithCustomDefaultTemplate)
-
-	_, stderr, err := helpers.RunTerragruntCommandWithOutput(t, fmt.Sprintf(
-		"terragrunt --non-interactive --working-dir %s scaffold %s",
-		filepath.Join(testPath, "unit"),
-		testScaffoldModuleURL,
-	))
-
-	require.NoError(t, err)
-	assert.Contains(t, stderr, "Scaffolding completed")
-
-	assert.FileExists(t, filepath.Join(testPath, "unit", "terragrunt.hcl"))
-	assert.FileExists(t, filepath.Join(testPath, "unit", "external-template.txt"))
 }
